[content] feat: add 5 new cyberpunk quests for Montreal and Miami
- Quebec French (Montreal 2020-2029): Linguistic preservation against corporate Neo-English
- Drug Trafficking (Miami 2020-2029): Pharmaceutical cartel wars and synthetic addiction
- Walls of Wynwood (Miami 2020-2029): Street art resistance against corporate gentrification
- Miami Heat Game (Miami 2020-2029): Basketball team fight against cybernetic sports
- Alligators Everglades (Miami 2020-2029): Environmental mutation and wetland preservation

Related Issues: #140927007, #140927104, #140927108, #140927111, #140927113

+++ services/reset-service-go-service-go/server/qa_integration_test.go
@@ -148,7 +148,7 @@ func TestGetResetStatsQA(t *testing.T) {
 	assert.Equal(t, 42, resp.TotalResets)
 	assert.Equal(t, 38, resp.SuccessfulResets)
 	assert.Equal(t, 4, resp.FailedResets)
-	assert.Equal(t, 45.5, resp.AverageCompletionTime)
+	assert.Equal(t, float32(45.5), resp.AverageCompletionTime)
 }
 
 // TestTriggerResetQA tests reset triggering with various scenarios
